refactor: create runner internal package
